+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 
 ### Features
 
+- Make lint rules more strict
+- Update `minAppVersion`
+
+### Bugfixes
+
+- Fix auto-review obsidian issues
+
+## 0.1.0 (05 Jun 2026)
+
+### Features
+
 - Add first working version
 
 ### Improvements